# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GeoPagos Reto Api. Edwin Jose Mogollon Escalona eJme1992",
        "description": "Documentación de la API para la gestión de torneos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/genders/all": {
            "get": {
                "tags": [
                    "Genders"
                ],
                "summary": "Obtener todos los géneros",
                "description": "Devuelve una lista de todos los géneros con los campos `slug` y `name`. Requiere un token JWT para la autenticación.",
                "operationId": "1f169fd9e839946d4eba1e327918c485",
                "responses": {
                    "200": {
                        "description": "Lista de géneros obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "slug": {
                                                "type": "string",
                                                "example": "femenino"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Femenino"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener la lista de géneros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}